+++ RELEASENOTES.md
@@ -17,10 +17,16 @@
 
 9. New API for orders tracking url GET /orders/track/{orderId}
 
+10. Remove https:// for image url in product, product_assets, store_assets
+
 ##DB Changes
 ALTER TABLE voucher ADD requireToClaim TINYINT(1) NOT NULL DEFAULT 1;
 ALTER TABLE customer_voucher ADD guestEmail VARCHAR(100);
 ALTER TABLE customer_voucher ADD guestVoucher TINYINT(1) NOT NULL DEFAULT 0;
+
+##Config changes:
+asset.service.URL=https://assets.symplified.it
+
 
 ##################################################
 # order-service-3.8.2-SNAPSHOT |24-June-2022